Refactored the blogExplorer file and added unit tests.
